Rename directory
Rename sub-directory for chart controls for extra indicator-specific parameters.
