+++ src/app/app.module.ts
@@ -41,7 +41,7 @@ import { LoginComponent } from './login/login.component';
 import { NavbarComponent } from './navbar/navbar.component';
 import { PageNotFoundComponent } from './http-status/page-not-found/page-not-found.component';
 import { SidebarComponent } from './sidebar/sidebar.component';
-import { ThresholdComponent } from './charts/extra-components/threshold.component';
+import { ThresholdComponent } from './charts/extra-params-components/threshold.component';
 import { WaveComponent } from './ng2-spin-kit/wave.component';
 
 // App services
