+++ src/app/charts/extra-params-components/threshold.component.ts
